refactor(teams): split MembersPage into focused modules
