+++ src/components/members/MembersPage.tsx
@@ -3,795 +3,35 @@ import { Icon } from "@iconify/react";
 import { useTranslation } from "react-i18next";
 import { useVaultStore } from "@/stores/vaultStore";
 import { useTeamStore } from "@/stores/teamStore";
-import type { TeamMember, TeamRole } from "@/stores/teamStore";
+import type { TeamMember } from "@/stores/teamStore";
 import { useSubscriptionStore } from "@/stores/subscriptionStore";
 import { useUIStore } from "@/stores/uiStore";
 import { useTeamSessionStore } from "@/stores/teamSessionStore";
 import { useHistoryStore } from "@/stores/historyStore";
-import { MiniAvatar, avatarColor } from "@/components/shared/AvatarStack";
-import { PresenceAvatar } from "@/components/shared/PresenceAvatar";
-import { UserSearchField } from "@/components/shared/UserSearchField";
+import { avatarColor } from "@/components/shared/AvatarStack";
 import { getMyUserId } from "@/services/teamService";
 import { getMyHandle } from "@/services/account";
-import { BaseCard } from "@/components/shared/BaseCard";
 import type { ContextMenuItem } from "@/components/shared/ContextMenu";
 import { SidePanelLayout } from "@/components/shared/SidePanelLayout";
 import { DragSelectSurface } from "@/components/shared/DragSelectSurface";
-import { ToolbarViewControls } from "@/components/shared/ToolbarViewControls";
-import type { LayoutMode, SortMode } from "@/components/shared/ToolbarViewControls";
-import { PanelShell, PanelHeader, PanelHeaderIconButton, FormSection } from "@/components/shared/Panel";
+import { PanelShell, PanelHeader, PanelHeaderIconButton } from "@/components/shared/Panel";
 import { useDragSelection } from "@/hooks/useDragSelection";
 import { useListKeyNav } from "@/hooks/useListKeyNav";
-import BuySeatsModal from "@/components/settings/BuySeatsModal";
 import { effectivePermissions, hasBuiltinRole, PERM_BITS } from "@/hooks/usePermission";
 import { runTeamAction } from "@/services/teamActionFeedback";
-import { openBillingCheckout } from "@/services/billingCheckout";
-import { RoleModal, TeamRolesPanel } from "@/components/settings/sections/RolesSection";
-import { seatAvailability } from "@/services/seatMath";
+import { TeamRolesPanel } from "@/components/settings/sections/RolesSection";
 import { guestCapFor, inviteSessionOf, memberHasAccess, seatUsage, sessionDisplayName } from "@/services/teamSharing";
-import { SeatsMeter } from "@/components/members/SeatsMeter";
-import { ROLE_META, RoleToggleChip, RolePermissionTooltip, RoleBlurb } from "@/components/members/roleChips";
-import { useUserSearch, type UserSearchResult } from "@/hooks/useUserSearch";
-import { inviteUserWithRoles, inviteByEmailAddress, inviteFailureReason, removeTeamMember, revokeInvitation } from "@/services/vaultShare";
+import { RoleToggleChip } from "@/components/members/roleChips";
 import { ConvertToTeamGate } from "@/components/vault-share/ConvertToTeamGate";
-import { assignableRoles, leastPrivilegedRole } from "@/components/vault-share/vaultShareModel";
 import { OffboardingDialog } from "@/components/members/OffboardingDialog";
-import type { DepartMode } from "@/services/teamOffboarding";
 import { PendingInviteCard } from "@/components/members/cards/PendingInviteCard";
-
-function RoleChip({ role }: { role: TeamRole }) {
-  const [showTip, setShowTip] = useState(false);
-  const meta = ROLE_META[role.name];
-  const color = role.color ?? meta?.color ?? avatarColor(role.name);
-  const bg = meta?.bg ?? `${color}1a`;
-
-  return (
-    <span className="relative inline-flex items-center" style={{ verticalAlign: "middle" }}>
-      <span
-        className="inline-flex items-center gap-1 text-[10px] font-medium px-2 py-0.5 rounded-full capitalize shrink-0 cursor-default"
-        style={{ color, background: bg }}
-        onMouseEnter={() => setShowTip(true)}
-        onMouseLeave={() => setShowTip(false)}
-      >
-        {role.is_builtin
-          ? <Icon icon="lucide:lock" width={8} style={{ opacity: 0.6 }} />
-          : <Icon icon="lucide:sparkles" width={8} style={{ opacity: 0.7 }} />
-        }
-        {role.name}
-      </span>
-      {showTip && <RolePermissionTooltip role={role} color={color} />}
-    </span>
-  );
-}
-
-function RoleBadges({
-  member, roles, canManage, onAddRole,
-}: {
-  member: TeamMember;
-  roles: TeamRole[];
-  canManage?: boolean;
-  onAddRole?: () => void;
-}) {
-  const { t } = useTranslation();
-  const memberRoles = member.role_ids
-    .map((rid) => roles.find((r) => r.id === rid))
-    .filter(Boolean) as TeamRole[];
-  memberRoles.sort((a, b) => a.position - b.position);
-  if (memberRoles.length === 0) {
-    if (canManage && onAddRole) {
-      return (
-        <button
-          onClick={(e) => { e.stopPropagation(); onAddRole(); }}
-          className="text-[10px] transition-colors"
-          style={{ color: "var(--t-text-dim)" }}
-          onMouseEnter={(e) => { (e.currentTarget as HTMLButtonElement).style.color = "var(--t-accent)"; }}
-          onMouseLeave={(e) => { (e.currentTarget as HTMLButtonElement).style.color = "var(--t-text-dim)"; }}
-        >
-          {t("members.noRoleAddPrompt")}
-        </button>
-      );
-    }
-    return <span className="text-[10px] text-(--t-text-dim)">{t("members.noRole")}</span>;
-  }
-  return (
-    <div className="flex flex-wrap gap-1">
-      {memberRoles.map((r) => <RoleChip key={r.id} role={r} />)}
-    </div>
-  );
-}
-
-// ─── Toolbar ──────────────────────────────────────────────────────────────────
-
-interface MembersToolbarProps {
-  search: string;
-  onSearchChange: (v: string) => void;
-  layoutMode: LayoutMode;
-  onLayoutModeChange: (v: LayoutMode) => void;
-  sortMode: SortMode;
-  onSortModeChange: (v: SortMode) => void;
-  canInvite: boolean;
-  showInvitePanel: boolean;
-  onToggleInvite: () => void;
-  pendingCount?: number;
-  canManageRoles?: boolean;
-  showRolesPanel?: boolean;
-  onToggleRoles?: () => void;
-  selectedCount: number;
-  vaultTabs?: { id: string; name: string }[];
-  primaryVaultId: string | null;
-  onSelectVault: (id: string) => void;
-}
-
-function MembersToolbar({
-  search, onSearchChange,
-  layoutMode, onLayoutModeChange,
-  sortMode, onSortModeChange,
-  canInvite, showInvitePanel, onToggleInvite,
-  pendingCount,
-  canManageRoles, showRolesPanel, onToggleRoles,
-  selectedCount,
-  vaultTabs, primaryVaultId, onSelectVault,
-}: MembersToolbarProps) {
-  const { t } = useTranslation();
-  return (
-    <div
-      className="flex items-center gap-2 px-5 py-2.5 shrink-0"
-      style={{ borderBottom: "1px solid var(--t-border)", background: "var(--t-bg-toolbar)" }}
-    >
-      <div className="flex items-center gap-2 min-w-0">
-        {vaultTabs && vaultTabs.length > 1 && (
-          <div className="flex items-center gap-1 shrink-0">
-            {vaultTabs.map(({ id, name }) => (
-              <button
-                key={id}
-                onClick={() => onSelectVault(id)}
-                className="px-2.5 py-1 rounded-lg text-xs font-medium transition-colors"
-                style={{
-                  background: id === primaryVaultId ? "var(--t-accent)" : "var(--t-bg-elevated)",
-                  color: id === primaryVaultId ? "#fff" : "var(--t-text-dim)",
-                }}
-              >
-                {name}
-              </button>
-            ))}
-          </div>
-        )}
-
-        <ToolbarViewControls
-          search={search}
-          onSearchChange={onSearchChange}
-          filterPlaceholder={t("members.toolbar.filterPlaceholder")}
-          filterShortcutId="filter"
-          layoutMode={layoutMode}
-          onLayoutModeChange={onLayoutModeChange}
-          sortMode={sortMode}
-          onSortModeChange={onSortModeChange}
-          extraSortOptions={[{ value: "role-asc", label: t("members.toolbar.sortByRole"), icon: "lucide:shield" }]}
-          filterWidth={176}
-        />
-      </div>
-
-      <div className="ml-auto flex items-center gap-2 shrink-0">
-        {selectedCount > 1 && (
-          <span className="text-xs text-(--t-text-dim) shrink-0">
-            {t("members.toolbar.selectedCount", { count: selectedCount })}
-          </span>
-        )}
-
-        {canManageRoles && onToggleRoles && (
-          <button
-            onClick={onToggleRoles}
-            className="flex items-center gap-1.5 px-3 py-1.5 rounded-lg text-xs font-medium transition-colors shrink-0"
-            style={{
-              background: showRolesPanel ? "color-mix(in srgb, var(--t-accent) 15%, transparent)" : "var(--t-bg-elevated)",
-              color: showRolesPanel ? "var(--t-accent)" : "var(--t-text-primary)",
-              border: `1px solid ${showRolesPanel ? "var(--t-accent)" : "var(--t-border)"}`,
-            }}
-          >
-            <Icon icon="lucide:shield" width={13} />
-            {t("members.roles")}
-          </button>
-        )}
-
-        {canInvite && (
-          <>
-            <div className="w-px h-5 self-center bg-(--t-border-hover)" />
-            <button
-              onClick={onToggleInvite}
-              className="relative flex items-center gap-1.5 px-3 py-1.5 rounded-lg text-xs font-medium transition-colors shrink-0"
-              style={{
-                background: showInvitePanel ? "var(--t-accent-hover)" : "var(--t-accent)",
-                color: "var(--t-on-accent, #fff)",
-                border: "1px solid var(--t-accent-hover)",
-              }}
-              onMouseEnter={(e) => (e.currentTarget.style.background = "var(--t-accent-hover)")}
-              onMouseLeave={(e) => (e.currentTarget.style.background = showInvitePanel ? "var(--t-accent-hover)" : "var(--t-accent)")}
-            >
-              <Icon icon="lucide:user-plus" width={13} />
-              {t("members.toolbar.inviteBtn")}
-              {!!pendingCount && (
-                <span
-                  className="absolute -top-1.5 -right-1.5 flex items-center justify-center text-[9px] font-bold rounded-full min-w-[16px] h-4 px-0.5"
-                  style={{ background: "var(--t-status-error)", color: "#fff" }}
-                >
-                  {pendingCount}
-                </span>
-              )}
-            </button>
-          </>
-        )}
-      </div>
-    </div>
-  );
-}
-
-// ─── Member card ──────────────────────────────────────────────────────────────
-
-interface MemberCardProps {
-  member: TeamMember;
-  roles: TeamRole[];
-  isMe: boolean;
-  isOwner: boolean;
-  isSelected: boolean;
-  isFocused: boolean;
-  layoutMode: LayoutMode;
-  canManage?: boolean;
-  onAddRole?: () => void;
-  onSelect: (id: string, e: React.MouseEvent<HTMLDivElement>) => void;
-  onDoubleClick: () => void;
-  contextMenuItems: ContextMenuItem[];
-  bulkContextMenuItems?: ContextMenuItem[];
-}
-
-function MemberAvatar({ member, size }: { member: TeamMember; size: number }) {
-  return <PresenceAvatar handle={member.handle} size={size} online={member.is_online} animate />;
-}
-
-function MemberCard({
-  member, roles, isMe, isOwner, isSelected, isFocused, layoutMode,
-  canManage, onAddRole,
-  onSelect, onDoubleClick, contextMenuItems, bulkContextMenuItems,
-}: MemberCardProps) {
-  const { t } = useTranslation();
-  if (layoutMode === "grid") {
-    return (
-      <BaseCard
-        data-selectable-id={member.user_id}
-        isSelected={isSelected}
-        isFocused={isFocused}
-        isList={false}
-        onClick={(e) => onSelect(member.user_id, e)}
-        onDoubleClick={onDoubleClick}
-        contextMenuItems={contextMenuItems}
-        bulkContextMenuItems={bulkContextMenuItems}
-        className="flex-col items-center text-center gap-2 py-4"
-      >
-        <div className="relative">
-          <MemberAvatar member={member} size={40} />
-          {isOwner && (
-            <span className="absolute -top-1.5 -right-1.5 flex items-center justify-center w-4 h-4 rounded-full" style={{ background: "rgba(167,139,250,0.2)", border: "1px solid rgba(167,139,250,0.35)" }}>
-              <Icon icon="lucide:crown" width={8} style={{ color: "#a78bfa" }} />
-            </span>
-          )}
-        </div>
-        <div className="w-full min-w-0 flex flex-col items-center gap-1">
-          <div className="flex items-center gap-1 justify-center">
-            <p className="text-xs font-medium truncate text-(--t-text-bright) max-w-[120px]">{member.handle}</p>
-            {isMe && (
-              <span className="text-[9px] px-1 py-0.5 rounded-sm shrink-0" style={{ color: "var(--t-text-dim)", background: "var(--t-bg-elevated)" }}>{t("members.youBadge")}</span>
-            )}
-          </div>
-          <RoleBadges member={member} roles={roles} canManage={canManage} onAddRole={onAddRole} />
-        </div>
-      </BaseCard>
-    );
-  }
-
-  return (
-    <BaseCard
-      data-selectable-id={member.user_id}
-      isSelected={isSelected}
-      isFocused={isFocused}
-      isList
-      onClick={(e) => onSelect(member.user_id, e)}
-      onDoubleClick={onDoubleClick}
-      contextMenuItems={contextMenuItems}
-      bulkContextMenuItems={bulkContextMenuItems}
-    >
-      <MemberAvatar member={member} size={32} />
-      <div className="flex-1 min-w-0">
-        <div className="flex items-center gap-1.5">
-          <p className="text-sm font-medium truncate text-(--t-text-bright)">{member.handle}</p>
-          {isOwner && <Icon icon="lucide:crown" width={11} style={{ color: "#a78bfa", flexShrink: 0 }} />}
-          {isMe && (
-            <span className="text-[10px] px-1.5 py-0.5 rounded-sm shrink-0" style={{ color: "var(--t-text-dim)", background: "var(--t-bg-elevated)" }}>{t("members.youBadge")}</span>
-          )}
-        </div>
-      </div>
-      <RoleBadges member={member} roles={roles} canManage={canManage} onAddRole={onAddRole} />
-    </BaseCard>
-  );
-}
-
-// ─── Member detail panel ──────────────────────────────────────────────────────
-
-interface MemberDetailPanelProps {
-  member: TeamMember;
-  isMe: boolean;
-  teamId: string;
-  teamRoles: TeamRole[];
-  canManageMembers: boolean;
-  isTargetOwner: boolean;
-  onClose: () => void;
-  onUpdated: () => void;
-}
-
-export function MemberDetailPanel({
-  member, isMe, teamId, teamRoles, canManageMembers, isTargetOwner, onClose, onUpdated,
-}: MemberDetailPanelProps) {
-  const { t } = useTranslation();
-  const assignMemberRole = useTeamStore((s) => s.assignMemberRole);
-  const removeMemberRole = useTeamStore((s) => s.removeMemberRole);
-  const push = useHistoryStore((s) => s.push);
-
-  const [error, setError] = useState("");
-  const [toggling, setToggling] = useState<string | null>(null);
-  const [justToggled, setJustToggled] = useState<string | null>(null);
-  const [offboarding, setOffboarding] = useState<DepartMode | null>(null);
-  const [creatingRole, setCreatingRole] = useState(false);
-
-  const canChangeRoles = canManageMembers && !isMe;
-  const canRemove = canManageMembers && !isTargetOwner && !isMe;
-  // The server rejects an owner removing themselves (teams.rs `is_owner`), so
-  // offering Leave to an owner would promise something that 403s.
-  const canLeave = isMe && !isTargetOwner;
-
-  const handleToggleRole = async (role: TeamRole) => {
-    const hasRole = member.role_ids.includes(role.id);
-    // Block removing the owner role from an owner
-    if (hasRole && isTargetOwner && role.is_builtin && role.name === "owner") {
-      setError(t("members.error.cannotRemoveOwnerRole"));
-      return;
-    }
-    setToggling(role.id);
-    setError("");
-    try {
-      if (hasRole) {
-        await runTeamAction({
-          pending: t("members.toast.removingRoleFrom", { role: role.name, name: member.handle }),
-          success: t("members.toast.roleRemovedFrom", { role: role.name, name: member.handle }),
-          run: () => removeMemberRole(teamId, member.user_id, role.id),
-        });
-        push({
-          label: t("members.history.removeRole", { name: member.handle }),
-          undo: async () => {
-            await useTeamStore.getState().assignMemberRole(teamId, member.user_id, role.id);
-            onUpdated();
-          },
-          redo: async () => {
-            await useTeamStore.getState().removeMemberRole(teamId, member.user_id, role.id);
-            onUpdated();
-          },
-        });
-      } else {
-        await runTeamAction({
-          pending: t("members.toast.assigningRoleTo", { role: role.name, name: member.handle }),
-          success: t("members.toast.roleAssignedTo", { role: role.name, name: member.handle }),
-          run: () => assignMemberRole(teamId, member.user_id, role.id),
-        });
-        push({
-          label: t("members.history.assignRole", { name: member.handle }),
-          undo: async () => {
-            await useTeamStore.getState().removeMemberRole(teamId, member.user_id, role.id);
-            onUpdated();
-          },
-          redo: async () => {
-            await useTeamStore.getState().assignMemberRole(teamId, member.user_id, role.id);
-            onUpdated();
-          },
-        });
-      }
-      onUpdated();
-      setJustToggled(role.id);
-      setTimeout(() => setJustToggled(null), 700);
-    } catch (e) {
-      setError(e instanceof Error ? e.message : t("members.error.failedToUpdateRole"));
-    } finally {
-      setToggling(null);
-    }
-  };
-
-  const joinedDate = new Date(member.joined_at).toLocaleDateString(undefined, {
-    year: "numeric", month: "long", day: "numeric",
-  });
-
-  return (
-    <>
-      {creatingRole && (
-        <RoleModal
-          teamId={teamId}
-          role={null}
-          onClose={() => { setCreatingRole(false); onUpdated(); }}
-        />
-      )}
-    <PanelShell>
-      <PanelHeader
-        icon="lucide:user"
-        title={member.handle ?? "?"}
-        subtitle={<RoleBadges member={member} roles={teamRoles} />}
-        onClose={onClose}
-      />
-
-      <div className="flex-1 overflow-y-auto p-4 space-y-3">
-        {/* Roles */}
-        <FormSection label={t("members.roles")}>
-          {canChangeRoles ? (
-            <div className="flex flex-wrap items-start gap-2">
-              {[...teamRoles]
-                .filter((r) => !(r.is_builtin && r.name === "owner"))
-                .sort((a, b) => a.position - b.position).map((role) => {
-
-                const hasRole = member.role_ids.includes(role.id);
-                const meta = ROLE_META[role.name];
-                const color = role.color ?? meta?.color ?? "var(--t-accent)";
-                const bg = meta?.bg ?? `${color}1a`;
-                return (
-                  <div key={role.id} className="flex flex-col gap-1">
-                  <button
-                    onClick={() => void handleToggleRole(role)}
-                    disabled={toggling === role.id}
-                    className="flex items-center gap-1.5 px-3 py-1.5 rounded-lg text-xs font-medium transition-all"
-                    style={{
-                      background: justToggled === role.id ? "rgba(52,211,153,0.15)" : hasRole ? bg : "var(--t-bg-elevated)",
-                      color: justToggled === role.id ? "#34d399" : hasRole ? color : "var(--t-text-dim)",
-                      border: `1px solid ${justToggled === role.id ? "#34d39944" : hasRole ? `${color}44` : "var(--t-border)"}`,
-                      opacity: toggling === role.id ? 0.6 : 1,
-                      transition: "background 0.3s, color 0.3s, border-color 0.3s",
-                    }}
-                  >
-                    {toggling === role.id
-                      ? <Icon icon="lucide:loader-circle" width={10} className="animate-spin" />
-                      : justToggled === role.id
-                        ? <Icon icon="lucide:check-check" width={10} />
-                        : hasRole
-                          ? <Icon icon="lucide:check" width={10} />
-                          : null
-                    }
-                    {role.name}
-                    {role.is_builtin
-                      ? <Icon icon="lucide:lock" width={9} style={{ color: "var(--t-text-dim)", opacity: 0.6 }} />
-                      : <Icon icon="lucide:sparkles" width={9} style={{ color: "var(--t-text-dim)", opacity: 0.7 }} />
-                    }
-                  </button>
-                  {hasRole && <RoleBlurb name={role.name} />}
-                  </div>
-                );
-              })}
-              <button
-                onClick={() => setCreatingRole(true)}
-                className="flex items-center gap-1 px-3 py-1.5 rounded-lg text-xs font-medium transition-colors"
-                style={{ color: "var(--t-accent)", border: "1px dashed var(--t-accent)", background: "transparent", opacity: 0.7 }}
-                onMouseEnter={(e) => { (e.currentTarget as HTMLButtonElement).style.opacity = "1"; }}
-                onMouseLeave={(e) => { (e.currentTarget as HTMLButtonElement).style.opacity = "0.7"; }}
-              >
-                <Icon icon="lucide:plus" width={10} />
-                {t("members.newRole")}
-              </button>
-            </div>
-          ) : (
-            <RoleBadges member={member} roles={teamRoles} />
-          )}
-        </FormSection>
-
-        {/* Info */}
-        <FormSection label={t("members.info")}>
-          <div className="space-y-2 text-xs">
-            <div className="flex items-center justify-between">
-              <span className="text-(--t-text-dim)">{t("members.memberSince")}</span>
-              <span className="text-(--t-text-primary)">{joinedDate}</span>
-            </div>
-            {member.invited_by_display_name && (
-              <div className="flex items-center justify-between gap-4">
-                <span className="text-(--t-text-dim) shrink-0">{t("members.invitedBy")}</span>
-                <span className="text-(--t-text-primary) truncate">{member.invited_by_display_name}</span>
-              </div>
-            )}
-          </div>
-        </FormSection>
-
-        {/* Danger zone */}
-        {(canRemove || canLeave) && (
-          <FormSection label={t("members.dangerZone")}>
-            <button
-              onClick={() => setOffboarding(canLeave ? "leave" : "remove")}
-              className="w-full flex items-center justify-center gap-2 px-3 py-2 rounded-lg text-xs font-medium transition-colors"
-              style={{
-                background: "var(--t-bg-elevated)",
-                color: "var(--t-status-error)",
-                border: "1px solid rgba(239,68,68,0.3)",
-              }}
-              onMouseEnter={(e) => { e.currentTarget.style.background = "rgba(239,68,68,0.08)"; }}
-              onMouseLeave={(e) => { e.currentTarget.style.background = "var(--t-bg-elevated)"; }}
-            >
-              <Icon icon={canLeave ? "lucide:log-out" : "lucide:user-minus"} width={13} />
-              {canLeave ? t("members.leaveTeam") : t("members.removeFromTeam")}
-            </button>
-          </FormSection>
-        )}
-
-        {error && <p className="text-xs px-1" style={{ color: "var(--t-status-error)" }}>{error}</p>}
-      </div>
-    </PanelShell>
-
-    {offboarding && (
-      <OffboardingDialog
-        members={[member]}
-        teamId={teamId}
-        mode={offboarding}
-        onClose={() => setOffboarding(null)}
-        onDone={() => { onClose(); onUpdated(); }}
-      />
-    )}
-    </>
-  );
-}
-
-// ─── Invite panel ─────────────────────────────────────────────────────────────
-
-function isValidEmail(s: string) {
-  return /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(s);
-}
-
-interface InvitePanelProps {
-  teamId: string;
-  existingIds: Set<string>;
-  teamRoles: TeamRole[];
-  onClose: () => void;
-  onMemberAdded: () => void;
-}
-
-export function InvitePanel({ teamId, existingIds, teamRoles, onClose, onMemberAdded }: InvitePanelProps) {
-  const { t } = useTranslation();
-  const { usedSeats, totalSeats, load: reloadSubscription } = useSubscriptionStore();
-  const { query, setQuery, results, searching, open, setOpen, inputRef, dropdownRef, reset } =
-    useUserSearch(existingIds);
-  const [selectedRoleIds, setSelectedRoleIds] = useState<string[]>([]);
-  const [adding, setAdding] = useState<string | null>(null);
-  const [sendingInvite, setSendingInvite] = useState(false);
-  const [success, setSuccess] = useState("");
-  const [error, setError] = useState("");
-  const [buySeatsFor, setBuySeatsFor] = useState<UserSearchResult | null | undefined>(undefined);
-
-  const { atLimit: isAtSeatLimit } = seatAvailability(usedSeats, totalSeats);
-
-  const builtinRoles = useMemo(() => assignableRoles(teamRoles), [teamRoles]);
-  const defaultMemberRoleId = useMemo(
-    () => builtinRoles.find((r) => r.is_builtin && r.name === "member")?.id,
-    [builtinRoles],
-  );
-  useEffect(() => {
-    if (defaultMemberRoleId && selectedRoleIds.length === 0) {
-      setSelectedRoleIds([defaultMemberRoleId]);
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [defaultMemberRoleId]);
-
-  const toggleRole = (roleId: string) =>
-    setSelectedRoleIds((prev) =>
-      prev.includes(roleId) ? prev.filter((id) => id !== roleId) : [...prev, roleId],
-    );
-  const hasRoleSelected = selectedRoleIds.length > 0;
-  // Never relied on by the UI (the invite actions are disabled without a selection) —
-  // a pure safety net so no future caller can end up granting "member" by accident.
-  const fallbackRoleName = useMemo(() => leastPrivilegedRole(teamRoles)?.name ?? "connect-only", [teamRoles]);
-  const primaryRoleName = useMemo(
-    () => builtinRoles.find((r) => selectedRoleIds.includes(r.id))?.name ?? fallbackRoleName,
-    [selectedRoleIds, builtinRoles, fallbackRoleName],
-  );
-  const selectedRoleLabel = useMemo(() => {
-    const names = selectedRoleIds
-      .map((id) => builtinRoles.find((r) => r.id === id)?.name)
-      .filter(Boolean);
-    return names.length > 0 ? names.join(", ") : t("members.invite.noRoleFallback");
-  }, [selectedRoleIds, builtinRoles, t]);
-
-  useEffect(() => { void reloadSubscription(); }, []);
-  useEffect(() => { inputRef.current?.focus(); }, []);
-
-  const handleAdd = async (user: UserSearchResult) => {
-    if (!hasRoleSelected) return;
-    if (isAtSeatLimit) { setBuySeatsFor(user); setOpen(false); return; }
-    setAdding(user.user_id); setError(""); setSuccess("");
-    try {
-      const result = await inviteUserWithRoles({
-        teamId,
-        userId: user.user_id,
-        handle: user.handle,
-        roleIds: selectedRoleIds,
-        roles: teamRoles,
-      });
-      reset();
-      setSuccess(result.status === "pending"
-        ? t("members.toast.invitationSentToUser", { name: user.handle })
-        : t("members.toast.userAdded", { name: user.handle }));
-      await reloadSubscription();
-      onMemberAdded();
-    } catch (e) {
-      const err = e instanceof Error ? e : new Error(String(e));
-      if ((e as { code?: number }).code === 402 || err.message.includes("402")) {
-        setBuySeatsFor(user); setOpen(false);
-      } else {
-        setOpen(false);
-        setError(t("members.error.inviteFailed", { name: user.handle, reason: inviteFailureReason(err) }));
-      }
-    } finally { setAdding(null); }
-  };
-
-  const handleEmailInvite = async () => {
-    if (!isValidEmail(query) || !hasRoleSelected) return;
-    if (isAtSeatLimit) { setBuySeatsFor(null); return; }
-    setSendingInvite(true); setError(""); setSuccess("");
-    try {
-      const invitedEmail = query;
-      await inviteByEmailAddress({ teamId, email: invitedEmail, roleName: primaryRoleName });
-      reset();
-      setSuccess(t("members.toast.invitationSentToEmail", { email: invitedEmail }));
-      await reloadSubscription();
-      onMemberAdded();
-    } catch (e) {
-      const err = e instanceof Error ? e : new Error(String(e));
-      if ((e as { code?: number }).code === 402 || err.message.includes("402")) {
-        setBuySeatsFor(null);
-      } else {
-        setOpen(false);
-        setError(t("members.error.inviteFailed", { name: query, reason: inviteFailureReason(err) }));
-      }
-    } finally { setSendingInvite(false); }
-  };
-
-  const showEmailInviteOption = open && results.length === 0 && !searching && isValidEmail(query);
-
-  return (
-    <>
-      {buySeatsFor !== undefined && (
-        <BuySeatsModal
-          teamId={teamId}
-          pendingUser={buySeatsFor ?? null}
-          pendingRole={primaryRoleName}
-          onClose={() => setBuySeatsFor(undefined)}
-          onSuccess={async () => {
-            setBuySeatsFor(undefined);
-            await reloadSubscription();
-            onMemberAdded();
-          }}
-        />
-      )}
-
-      <PanelShell>
-        <PanelHeader
-          icon="lucide:user-plus"
-          title={t("members.invite.title")}
-          onClose={onClose}
-        />
-
-        <div className="flex-1 overflow-y-auto p-4 space-y-4">
-          {/* Seats usage */}
-          <FormSection label={t("members.invite.seatsLabel")}>
-            <SeatsMeter onBuySeats={() => setBuySeatsFor(null)} />
-          </FormSection>
-
-          {/* Role selector */}
-          <FormSection label={t("members.invite.initialRoles")}>
-            <div className="flex flex-wrap items-start gap-2">
-              {builtinRoles.map((r) => (
-                <RoleToggleChip
-                  key={r.id}
-                  name={r.name}
-                  active={selectedRoleIds.includes(r.id)}
-                  onClick={() => toggleRole(r.id)}
-                  blurb
-                />
-              ))}
-            </div>
-            {!hasRoleSelected && (
-              <p className="text-xs mt-1.5" style={{ color: "var(--t-text-dim)" }}>{t("members.invite.selectRoleHint")}</p>
-            )}
-          </FormSection>
-
-          {/* Search input */}
-          <FormSection label={t("members.invite.searchOrInviteLabel")} className="overflow-visible">
-            <UserSearchField
-              placeholder={t("members.invite.searchUserPlaceholder")}
-              query={query}
-              onQueryChange={(v) => { setQuery(v); setSuccess(""); }}
-              onClear={() => { reset(); setSuccess(""); }}
-              onSubmitQuery={() => { if (isValidEmail(query) && results.length === 0) void handleEmailInvite(); }}
-              results={results}
-              searching={searching}
-              open={open}
-              setOpen={setOpen}
-              inputRef={inputRef}
-              dropdownRef={dropdownRef}
-              adding={adding}
-              addLabel={t("members.invite.addWithRole", { role: selectedRoleLabel })}
-              onAdd={(user) => void handleAdd(user)}
-              actionsDisabled={!hasRoleSelected}
-              emailOption={{
-                visible: showEmailInviteOption,
-                label: <>{t("members.invite.sendInviteLabel")} <span className="font-medium">{query}</span></>,
-                actionLabel: t("members.invite.inviteArrow"),
-                sending: sendingInvite,
-                onInvite: () => void handleEmailInvite(),
-              }}
-            />
-          </FormSection>
-
-          {error && <p className="text-xs px-1" style={{ color: "var(--t-status-error)" }}>{error}</p>}
-          {success && (
-            <div className="flex items-center gap-2 px-3 py-2.5 rounded-xl" style={{ background: "rgba(52,211,153,0.08)", border: "1px solid rgba(52,211,153,0.2)" }}>
-              <Icon icon="lucide:circle-check" width={14} style={{ color: "#34d399" }} />
-              <p className="text-xs" style={{ color: "#34d399" }}>{success}</p>
-            </div>
-          )}
-        </div>
-      </PanelShell>
-    </>
-  );
-}
-
-// ─── Upgrade CTA ──────────────────────────────────────────────────────────────
-
-function SignInToCloudCTA({ onSignIn }: { onSignIn: () => void }) {
-  const { t } = useTranslation();
-  return (
-    <div className="flex flex-col items-center justify-center h-full min-h-[320px] gap-5">
-      <div className="w-16 h-16 rounded-2xl flex items-center justify-center" style={{ background: "rgba(99,102,241,0.1)", border: "1px solid rgba(99,102,241,0.2)" }}>
-        <Icon icon="lucide:cloud" width={28} style={{ color: "var(--t-accent)" }} />
-      </div>
-      <div className="text-center">
-        <p className="text-sm font-medium mb-1 text-(--t-text-primary)">{t("members.cta.signInTitle")}</p>
-        <p className="text-xs max-w-[240px] text-(--t-text-dim)">
-          {t("members.cta.signInDesc")}
-        </p>
-      </div>
-      <button
-        onClick={onSignIn}
-        className="flex items-center gap-2 px-4 py-2 rounded-lg text-sm font-medium text-white transition-opacity hover:opacity-90"
-        style={{ background: "var(--t-accent)" }}
-      >
-        {t("members.cta.signInBtn")}
-      </button>
-    </div>
-  );
-}
-
-function UpgradeToTeamsCTA() {
-  const { t } = useTranslation();
-  const openCheckout = async () => {
-    await openBillingCheckout("teams");
-  };
-
-  return (
-    <div className="flex flex-col items-center justify-center h-full min-h-[320px] gap-5">
-      <div className="w-16 h-16 rounded-2xl flex items-center justify-center" style={{ background: "rgba(99,102,241,0.1)", border: "1px solid rgba(99,102,241,0.2)" }}>
-        <Icon icon="lucide:users-round" width={28} style={{ color: "var(--t-accent)" }} />
-      </div>
-      <div className="text-center">
-        <p className="text-sm font-medium mb-1 text-(--t-text-primary)">{t("members.cta.upgradeTitle")}</p>
-        <p className="text-xs max-w-[220px] text-(--t-text-dim)">
-          {t("members.cta.upgradeDesc")}
-        </p>
-      </div>
-      <button
-        onClick={() => void openCheckout()}
-        className="flex items-center gap-2 px-4 py-2 rounded-lg text-sm font-medium text-white transition-opacity hover:opacity-90"
-        style={{ background: "var(--t-accent)" }}
-      >
-        {t("members.cta.upgradeBtn")}
-      </button>
-    </div>
-  );
-}
+import { MembersToolbar } from "@/components/members/MembersToolbar";
+import { RoleBadges } from "@/components/members/roleBadges";
+import { MemberCard } from "@/components/members/cards/MemberCard";
+import { SelfCard } from "@/components/members/cards/SelfCard";
+import { MemberDetailPanel } from "@/components/members/panels/MemberDetailPanel";
+import { InvitePanel } from "@/components/members/panels/InvitePanel";
+import { SignInToCloudCTA, UpgradeToTeamsCTA } from "@/components/members/panels/MembersCTA";
 
 // ─── Main page ────────────────────────────────────────────────────────────────
 
@@ -1234,37 +474,7 @@ const vaultTabs = selectedVaultIds.length > 1
               className={layoutMode === "grid" ? "grid gap-4" : "flex flex-col gap-1"}
               style={layoutMode === "grid" ? { gridTemplateColumns: "repeat(auto-fill, minmax(160px, 1fr))" } : undefined}
             >
-              {layoutMode === "grid" ? (
-                <BaseCard isList={false} className="flex-col items-center text-center gap-2 py-4">
-                  <MiniAvatar name={myHandle || "?"} size={40} />
-                  <div className="w-full min-w-0 flex flex-col items-center gap-1">
-                    <div className="flex items-center gap-1 justify-center">
-                      {myHandle === null
-                        ? <div className="h-3.5 w-20 rounded-sm animate-pulse" style={{ background: "var(--t-bg-elevated)" }} />
-                        : <p className="text-xs font-medium truncate text-(--t-text-bright) max-w-[120px]">{myHandle || t("members.you")}</p>
-                      }
-                      <span className="text-[9px] px-1 py-0.5 rounded-sm shrink-0" style={{ color: "var(--t-text-dim)", background: "var(--t-bg-elevated)" }}>{t("members.youBadge")}</span>
-                    </div>
-                    <span className="text-[10px] font-medium px-2 py-0.5 rounded-full" style={{ color: "#a78bfa", background: "rgba(167,139,250,0.12)" }}>{t("members.ownerRoleLabel")}</span>
-                  </div>
-                </BaseCard>
-              ) : (
-                <BaseCard isList>
-                  <MiniAvatar name={myHandle || "?"} size={32} />
-                  <div className="flex-1 min-w-0">
-                    <div className="flex items-center gap-1.5">
-                      {myHandle === null
-                        ? <div className="h-3.5 w-40 rounded-sm animate-pulse" style={{ background: "var(--t-bg-elevated)" }} />
-                        : <p className="text-sm font-medium truncate text-(--t-text-bright)">{myHandle || t("members.you")}</p>
-                      }
-                      <span className="text-[10px] px-1.5 py-0.5 rounded-sm shrink-0" style={{ color: "var(--t-text-dim)", background: "var(--t-bg-elevated)" }}>
-                        {t("members.youBadge")}
-                      </span>
-                    </div>
-                  </div>
-                  <span className="text-[10px] font-medium px-2 py-0.5 rounded-full" style={{ color: "#a78bfa", background: "rgba(167,139,250,0.12)" }}>{t("members.ownerRoleLabel")}</span>
-                </BaseCard>
-              )}
+              <SelfCard handle={myHandle} layoutMode={layoutMode} />
             </div>
           </div>
         </div>
